+++ Project/Assets/ML-Agents/Examples/GridWorld/Scripts/GridArea.cs
@@ -14,7 +14,7 @@ public class GridArea : MonoBehaviour
 
     public GameObject trueAgent;
 
-    IFloatProperties m_ResetParameters;
+    FloatPropertiesChannel m_ResetParameters;
 
     Camera m_AgentCam;