+++ com.unity.ml-agents/Editor/BehaviorParametersEditor.cs
@@ -2,6 +2,7 @@ using MLAgents.Sensors;
 using UnityEditor;
 using Barracuda;
 using MLAgents.Policies;
+using UnityEngine;
 
 namespace MLAgents.Editor
 {
@@ -22,27 +23,55 @@ namespace MLAgents.Editor
         {
             var so = serializedObject;
             so.Update();
+            bool needPolicyUpdate; // Whether the name, model, inference device, or BehaviorType changed.
 
             // Drawing the Behavior Parameters
+            EditorGUI.indentLevel++;
+            EditorGUI.BeginChangeCheck(); // global
+
             EditorGUI.BeginChangeCheck();
-            EditorGUI.indentLevel++;
-            EditorGUILayout.PropertyField(so.FindProperty("m_BehaviorName"));
-            EditorGUILayout.PropertyField(so.FindProperty("m_BrainParameters"), true);
-            EditorGUILayout.PropertyField(so.FindProperty("m_Model"), true);
-            EditorGUI.indentLevel++;
-            EditorGUILayout.PropertyField(so.FindProperty("m_InferenceDevice"), true);
-            EditorGUI.indentLevel--;
-            EditorGUILayout.PropertyField(so.FindProperty("m_BehaviorType"));
-            EditorGUILayout.PropertyField(so.FindProperty("TeamId"));
-            EditorGUILayout.PropertyField(so.FindProperty("m_UseChildSensors"), true);
-            // EditorGUILayout.PropertyField(serializedObject.FindProperty("m_Heuristic"), true);
-            EditorGUI.indentLevel--;
-            if (EditorGUI.EndChangeCheck())
             {
-                m_RequireReload = true;
+                EditorGUILayout.PropertyField(so.FindProperty("m_BehaviorName"));
             }
+            needPolicyUpdate = EditorGUI.EndChangeCheck();
+
+            EditorGUI.BeginDisabledGroup(Application.isPlaying);
+            {
+                EditorGUILayout.PropertyField(so.FindProperty("m_BrainParameters"), true);
+            }
+            EditorGUI.EndDisabledGroup();
+
+            EditorGUI.BeginChangeCheck();
+            {
+                EditorGUILayout.PropertyField(so.FindProperty("m_Model"), true);
+                EditorGUI.indentLevel++;
+                EditorGUILayout.PropertyField(so.FindProperty("m_InferenceDevice"), true);
+                EditorGUI.indentLevel--;
+            }
+            needPolicyUpdate = needPolicyUpdate || EditorGUI.EndChangeCheck();
+
+            EditorGUI.BeginChangeCheck();
+            {
+                EditorGUILayout.PropertyField(so.FindProperty("m_BehaviorType"));
+            }
+            needPolicyUpdate = needPolicyUpdate || EditorGUI.EndChangeCheck();
+
+            EditorGUILayout.PropertyField(so.FindProperty("TeamId"));
+            EditorGUI.BeginDisabledGroup(Application.isPlaying);
+            {
+                EditorGUILayout.PropertyField(so.FindProperty("m_UseChildSensors"), true);
+            }
+            EditorGUI.EndDisabledGroup();
+
+            EditorGUI.indentLevel--;
+            m_RequireReload = EditorGUI.EndChangeCheck();
             DisplayFailedModelChecks();
             so.ApplyModifiedProperties();
+
+            if (needPolicyUpdate)
+            {
+                UpdateAgentPolicy();
+            }
         }
 
         /// <summary>
@@ -85,6 +114,22 @@ namespace MLAgents.Editor
                         EditorGUILayout.HelpBox(check, MessageType.Warning);
                     }
                 }
+            }
+        }
+
+        void UpdateAgentPolicy()
+        {
+            if (Application.isPlaying)
+            {
+                var behaviorParameters = (BehaviorParameters)target;
+                var agent = behaviorParameters.GetComponent<Agent>();
+                if (agent == null)
+                {
+                    return;
+                }
+
+                agent.ReloadPolicy();
+
             }
         }
     }